Separated each of the classes into their own files.
Created PosterCollection and LinkCollection classes with easy access properties. Changed all instances of List<Poster> to PosterCollection.
LinkCollection still needs more work.

Added method to Tomato to return a Movie by the input Url so a Movie objects Link can be parsed by the Tomato class.

When a Movie object is created by the Tomato class it now automatically appends the apikey to each of the Link objects.