Changed LinkCollection to MovieLinkCollection.
Added MovieResultLinkCollection for MovieSearchResult Links.

Changed List<Link> in Movie and MovieSearchResult to MovieLinkCollection and MovieResultLinkCollection respectively. In MovieResultLinkColletion, if "Next" or "Previous" links are null it will now return the current page.

Updated Parser.cs to parse for Movie Links and for MovieSearchResult Links.
In Tomato.cs changed FindMovieByQuery to FindMoviesByQuery to reflect that it returns multiple movie objects. Added a method to add the Api Key to the end of the MovieSearchResult Links.

In MVC project, updated MovieSearch page to access thumbnail posters using "item.Posters.ThumbnailPoster" instead of "item.Posters[0].Url"